# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Abroad Matrimony API
  version: 1.0.0
  description: |
    REST API for the Abroad Matrimony matchmaking platform.

    **Authentication:** All user endpoints (except OTP request/verify and token refresh)
    require a Bearer access token issued by `POST /api/v1/auth/otp/verify` or
    `POST /api/v1/auth/token/refresh`.

    Admin endpoints require a Bearer admin token from `POST /admin/auth/login`.

    Webhook endpoints (`/stripe/webhook`, `/razorpay/webhook`) use provider signature
    verification instead of Bearer tokens.

    **Response envelope:** Every response follows `ApiResponse<T>`:
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": { "hasMore": false, "cursor": null },
      "requestId": "req_abc123"
    }
    ```
    Errors add `"error": { "code": "ERROR_CODE", "message": "Human-readable message" }`.

    **All money values** are integers in paise (INR) or cents (USD). Never floats.

  contact:
    name: Abroad Matrimony Engineering
  license:
    name: Private
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.abroadmatrimony.com
    description: Production

# ── Security Schemes ──────────────────────────────────────────────────────────
components:
  securitySchemes:
    userBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token issued by /api/v1/auth/otp/verify or /token/refresh. Expires in 15 minutes.
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT issued by /admin/auth/login. Longer-lived; carries AdminRole claim.

# ── Reusable Schemas ──────────────────────────────────────────────────────────
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:    { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: "Phone must be in E.164 format" }
        details: { type: object, additionalProperties: true }
...
    SavedProfileLabel:
      type: string
      enum: [HIGH_FIT, EXPIRING, MAYBE]
      description: User-applied label on a saved profile card.
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:   { type: boolean, enum: [false] }
        error:     { $ref: '#/components/schemas/ApiError' }
        requestId: { type: string }

  # ── Reusable Parameters ───────────────────────────────────────────────────
...
  responses:
    400BadRequest:
      description: Validation error or bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            success: false
            error: { code: VALIDATION_ERROR, message: "Phone must be in E.164 format" }
    401Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            success: false
            error: { code: UNAUTHORIZED, message: "Access token is invalid or expired" }
...
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    500InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }

# ── Tags ──────────────────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/saved:
    get:
      tags: [Saved Profiles]
      summary: List saved profiles
      operationId: listSavedProfiles
      security:
        - userBearerAuth: []
      parameters:
        - name: label
          in: query
          required: false
          schema: { $ref: '#/components/schemas/SavedProfileLabel' }
      responses:
        '200':
          description: Saved profile list
          content:
            application/json:
              example:
                success: true
                data:
                  - id: "save-uuid"
                    savedUserId: "usr-uuid-arjun"
                    label: "INTERESTED"
                    notes: "Loved his story prompt"
                    createdAt: "2026-05-29T08:00:00.000Z"
                    savedUser: { id: "usr-uuid-arjun", name: "Arjun", currentCity: "London", currentCountry: "UK", verificationStatus: "APPROVED" }
                meta: { total: 1 }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '500': { $ref: '#/components/responses/500InternalError' }

    post:
      tags: [Saved Profiles]
      summary: Save a profile
      operationId: saveProfile
      security:
        - userBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [savedUserId]
              properties:
                savedUserId: { type: string, format: uuid }
                label: { $ref: '#/components/schemas/SavedProfileLabel' }
                notes: { type: string, maxLength: 500 }
            example:
              savedUserId: "usr-uuid-arjun"
              label: "INTERESTED"
      responses:
        '201':
          description: Profile saved
          content:
            application/json:
              example:
                success: true
                data: { id: "save-uuid", savedUserId: "usr-uuid-arjun", label: "INTERESTED", notes: null, createdAt: "2026-05-29T08:00:00.000Z" }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '409':
          description: Profile already saved
          content:
            application/json:
              example:
                success: false
                error: { code: CONFLICT, message: "Profile already saved" }
        '500': { $ref: '#/components/responses/500InternalError' }
  /api/v1/saved/compare:
    get:
      tags: [Saved Profiles]
      summary: Compare 2–3 saved profiles side by side
      operationId: compareSavedProfiles
      security:
        - userBearerAuth: []
      parameters:
        - name: ids
          in: query
          required: true
          description: Comma-separated list of 2–3 saved user UUIDs
          schema: { type: string }
          example: "uuid-a,uuid-b"
      responses:
        '200':
          description: Comparison data for each profile
          content:
            application/json:
              example:
                success: true
                data:
                  - savedUserId: "usr-uuid-arjun"
                    label: "INTERESTED"
                    notes: null
                    profile: { name: "Arjun", dateOfBirth: "1995-03-15T00:00:00.000Z", gender: "MALE", currentCity: "London", currentCountry: "UK", settlementIntent: "SETTLE_ABROAD", bio: null, completionScore: 85, verificationStatus: "APPROVED", trustScore: 7 }
                    realLifeAnswers: [{ questionKey: "faith", answer: "important" }]
                meta: { total: 2 }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalError' }
  /api/v1/saved/{savedUserId}:
    patch:
      tags: [Saved Profiles]
      summary: Update label and/or notes on a saved profile
      operationId: updateSavedProfile
      security:
        - userBearerAuth: []
      parameters:
        - name: savedUserId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label: { $ref: '#/components/schemas/SavedProfileLabel' }
                notes: { type: string, maxLength: 500 }
            example:
              label: "MAYBE"
              notes: "Good match, revisit in a week"
      responses:
        '200':
          description: Saved profile updated
          content:
            application/json:
              example:
                success: true
                data: { id: "save-uuid", savedUserId: "usr-uuid-arjun", label: "MAYBE", notes: "Good match" }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalError' }

    delete:
      tags: [Saved Profiles]
      summary: Remove a profile from saved list
      operationId: unsaveProfile
      security:
        - userBearerAuth: []
      parameters:
        - name: savedUserId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Profile removed from saved list
          content:
            application/json:
              example:
                success: true
                data: null
                meta: { message: "Profile removed from saved" }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalError' }
...